feat: Update Area method in Demo1.cs
Refactor the Area method in the Demo1 class of the language-ext-example project to correctly calculate the area of a circle. The previous implementation used the wrong variable in the Square function, resulting in incorrect calculations. This change fixes the issue by using the correct variable and improves the accuracy of the area calculation.

Fixes #123